docs: recommend ECR instead of DockerHub in the README (#316)
*Issue #, if available:*

Recommend ECR instead of DockerHub as the source for prebuilt images. All else being equal, there's no reason for AWS customers to depend on a non-AWS product to receive AWS software when we have a comparable image registry of our own.

*Description of changes:*


By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.